+++ iris/service/content/petition/acl.py
@@ -11,3 +11,12 @@ class PublicPetitionServiceAuthFactory(BaseAuthFactory):
     __acl__ = [
         (security.Allow, acl.Roles.Admin, acl.Permissions.AdminPetition)
     ]
+
+
+class SupportersTokenServiceAuthFactory(BaseAuthFactory):
+    """Allows access with API key permission only
+    """
+
+    __acl__ = [
+        (security.Allow, acl.Roles.ApiKeyUser, acl.Permissions.ListSupporters)
+    ]